Enabled the '-Wextra' compiler setting with some relaxations.
This enables some extra warning flags that are not enabled by -Wall.
https://gcc.gnu.org/onlinedocs/gcc-4.9.3/gcc/Warning-Options.html#index-Wextra-285

Tested various settings on various combinations of compiler versions on
Debian and Raspbian systems, with different kernels and disabling
sign-compare and missing-field-initializers was needed as to not get
overblown by warnings (a lot from the kernel sources themselves).