Updated cfg80211.h's add_virtual_inf implementations with new parameter.
name_assign_type was added in kernel 4.1.
The original code in driver 4.3.22-beta already had that parameter added
to the cfg80211_rtw_add_virtual_intf function, it did not have it for
the rtw_cfg80211_add_monitor_if function. I did add it because
https://github.com/torvalds/linux/blob/master/drivers/staging/rtl8723au/os_dep/ioctl_cfg80211.c
does have it and it's also present in the 4.3.16 code branch.
So I applied the commit from
https://github.com/diederikdehaas/rtl8812AU/commit/92a3b52467d3f1e3a4e8894269405e8c142cdff9
also to this branch.