Fixed several 'misleading-indentation' warnings.
Compiling against Debian kernel 4.9.0-3-amd brought 2 types of
compilation warnings to light.
The 'misleading-indentation' are fixed with this commit.
There is also another one 'maybe-uninitialized' wrt 'EVM' variable in
hal/phydm/phydm_hwconfig.c line 1543, which remains open as I don't know
the proper way to fix it. It is caused by the 'if' construct on line
1512 not having an else clause and in that case 'EVM' is indeed
uninitialized.