Fixed 'switch-unreachable' and 'implicit-fallthrough' warnings.
Code in a switch statement, but outside a case clause will not be
executed. As it was always assigned a value in the following code, I
removed the initializing statement so it's just a variable declaration.

By adding an explict comment saying 'Intentional fallthrough' you hint
at the compiler that it isn't an error/omission, but you explicitly want
that (feature).